+++ src/main/java/ca/uvic/seng330/assn3/model/devices/Camera.java
@@ -1,10 +1,11 @@
 package ca.uvic.seng330.assn3.model.devices;
 
+import ca.uvic.seng330.assn3.logging.Logging;
 import ca.uvic.seng330.assn3.model.Hub;
-import ca.uvic.seng330.assn3.model.HubRegistrationException;
 import ca.uvic.seng330.assn3.model.IOEEventType;
 import java.util.UUID;
 import org.json.JSONObject;
+import org.slf4j.event.Level;
 
 public class Camera extends Device {
 
@@ -47,33 +48,25 @@ public class Camera extends Device {
       this.setStatus(Status.OFF);
     }
     this.diskSize = 0;
+    Logging.logWithID("Disk cleared.", getIdentifier(), Level.INFO);
   }
 
   public void record() throws CameraFullException {
     if (this.getStatus() == Status.ON) {
       if (this.isRecording) {
         this.diskSize++;
-        // TODO Logging
-        //      getMediator().log("Increasing Disk Size. Turning camera off.", Level.INFO,
-        // getIdentifier());
-        //    } else {
-        //      getMediator().log("Turning camera on.", Level.INFO, getIdentifier());
       }
-
       if (this.diskSize >= maxSize) {
         this.setStatus(Status.ERROR);
-        //    TODO  getMediator().log("Camera full", Level.WARN, getIdentifier());
-        try {
-          getHub().alert("Camera Full", this);
-        } catch (HubRegistrationException e) {
-          //    TODO	  getHub().log("Failed to Alert Hub. Camera has been unregistered", Level.WARN,
-          // getIdentifier());
-        }
+        Logging.logWithID("Camera Full.", getIdentifier(), Level.WARN);
+        getHub().alert("Camera Full", this); // TODO Do we still need this?
         throw new CameraFullException();
       }
       this.isRecording = !this.isRecording;
+      Logging.logWithID("Camera recording " + (isRecording ? "started" : "ended"),  getIdentifier(), Level.INFO);
     } else {
       this.isRecording = false;
+      Logging.logWithID("Camera turned off. Cannot record.", getIdentifier(), Level.WARN);
     }
   }
 
@@ -106,11 +99,9 @@ public class Camera extends Device {
     if (status == Status.OFF && this.isRecording()) {
       try {
         this.record();
-      } catch (CameraFullException e) {
-        // TODO Auto-generated catch block
-        e.printStackTrace();
-      }
+      } catch (CameraFullException e) {}
     }
     this.status = status;
+    Logging.logWithID("Camera turned " + status.toString(),  getIdentifier(), Level.INFO);
   }
 }